+++ tests/unit.test.ts
@@ -404,4 +404,11 @@ describe("Representation", () => {
 		expect(fn({ x: 1, y: 2 })).toEqual([ 1, 2, 0 ]);
 		expect(fn({ x: 1, y: 2, z: 3 })).toEqual([ 1, 2, 3 ]);
 	});
+	it("Converts an instance into primitive value", () => {
+		const a = vector([ 3, 4 ]);
+		const b = vector([ 6, 8 ]);
+
+		// @ts-expect-error: testing coercion
+		expect(a + b).toBe(15);
+	});
 });